refactor: split add/del/set code into files
